pyproject.toml
[tool.poetry]
name = "social-network-api"
version = "0.1.0"
description = ""
authors = ["jean.dossantos <jeandsantos88@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
databases = {extras = ["aiosqlite"], version = "^0.8.0"}
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
python-dotenv = "^1.0.0"
pydantic = "^2.5.2"
httpx = "^0.25.2"
pydantic-settings = "^2.1.0"
rich = "^13.7.0"
asgi-correlation-id = "^4.2.0"
python-json-logger = "^2.0.7"
logtail-python = "^0.2.7"
python-jose = "^3.3.0"
python-multipart = "^0.0.6"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.6"
black = "^23.11.0"
isort = "^5.12.0"
pylint = "^3.0.2"
pytest = "^7.4.3"
pytest-mock = "^3.12.0"


[tool.poetry.group.debug.dependencies]
ipykernel = "^6.27.1"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"